+++ gigaspatial/config.py
@@ -71,9 +71,7 @@ class Config(BaseSettings):
     GIGA_SCHOOL_MEASUREMENTS_API_KEY: str = Field(
         default="", alias="GIGA_SCHOOL_MEASUREMENTS_API_KEY"
     )
-    GIGA_COUNTRIES_API_KEY: str = Field(
-        default="", alias="GIGA_COUNTRIES_API_KEY"
-    )
+    GIGA_COUNTRIES_API_KEY: str = Field(default="", alias="GIGA_COUNTRIES_API_KEY")
 
     # Healthsites configuration
     # Used for accessing Healthsites global health facility data
@@ -149,6 +147,7 @@ class Config(BaseSettings):
             "cell_tower": "cell_tower",
             "cell": "cell",
             "mobile_coverage": "mobile_coverage",
+            "wri": "wri",
         },
         description="Mapping of data types to directory names",
     )